Normalise offsetX and offsetY on the event
According to the W3C Working Draft, offsetX and offsetY should be
relative to the padding edge of the target element. The only browser
using this convention is IE. Webkit uses the border edge, Opera uses the
content edge, and FireFox does not support the properties.

Normalizing to the border edge is easiest to do.